change test function name (chech -> check) refactor kind completion
update test
