+++ testdata/completion/Ordered/LocalBefore.scala
@@ -1,8 +1,9 @@
 class LocalBefore {
   val field1 = 45
-  val field2 = 33
 
   def fiFoo = ???
+
+  val fil1, fil2: Int
 
   def foo2(fiParam: Int) = {
     val fiValue = 67
@@ -13,7 +14,8 @@ class LocalBefore {
 fiParam
 fiValue
 field1
-field2
+fil1
+fil2
 fiFoo
 finalize
 final
